New hyp3-job service
