+++ src/app/services/hyp3/hyp3-job.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { Hyp3JobService } from './hyp3-job.service';
+
+describe('Hyp3JobService', () => {
+  let service: Hyp3JobService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(Hyp3JobService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
