+++ src/app/services/hyp3/index.ts
@@ -2,3 +2,4 @@ export * from './hyp3-api.service';
 export * from './on-demand.service';
 export * from './possible-hyp3-jobs.service';
 export * from './hyp3-job-status.service';
+export * from './hyp3-job.service';
